feat: separate service from gasket/io boilerplate
* use exclusively async IOs (this is bothersome as it introduces new
dependencies which I am not sure we want to pull in...)
* separate the gasket and IO boilerplate from the actual echo service
* I have not fully separated the pure business logic (echo) from the
node mgmt stuff (node_id)
