+++ Cargo.toml
@@ -19,6 +19,7 @@ async-trait = "0.1.83"
 bech32 = "0.11.0"
 clap = "4.5.20"
 ctor = "0.2"
+futures-util = "0.3.31"
 gasket = "0.8.0"
 hex = "0.4.3"
 indicatif = "0.17.9"
