+++ crates/amaru-sim/src/echo/run.rs
@@ -14,18 +14,14 @@
 
 use std::time::Duration;
 
-use super::message::{
-    Envelope,
-    Message::{self, *},
+use crate::echo::{
+    io::{DoEcho, ReadInput, SendOutput},
+    service::EchoService,
 };
-use gasket::{framework::*, runtime::Tether};
-use tokio::io::stdin;
-use tokio_stream::StreamExt;
-use tokio_util::{
-    codec::{FramedRead, LinesCodec},
-    sync::CancellationToken,
-};
-use tracing::{info, trace};
+
+use gasket::runtime::Tether;
+use tokio_util::sync::CancellationToken;
+use tracing::trace;
 
 pub async fn run() {
     let echo_pipeline = bootstrap();
@@ -47,7 +43,8 @@ pub async fn run() {
 ///    JSON.
 pub fn bootstrap() -> Vec<Tether> {
     let mut read_input = ReadInput::new();
-    let mut echo = DoEcho::new();
+    let service = Box::new(EchoService::new());
+    let mut echo = DoEcho::new(service);
     let mut send_output = SendOutput::new();
 
     let (to_echo, from_in) = gasket::messaging::tokio::mpsc_channel(50);
@@ -102,203 +99,5 @@ fn define_gasket_policy() -> gasket::runtime::Policy {
         bootstrap_retry: retries.clone(),
         work_retry: retries.clone(),
         teardown_retry: retries.clone(),
-    }
-}
-
-type EchoesOut = gasket::messaging::OutputPort<Envelope>;
-type EchoesIn = gasket::messaging::InputPort<Envelope>;
-
-struct Action {}
-
-struct ReadWorker {}
-
-#[derive(Stage)]
-#[stage(name = "read", unit = "Action", worker = "ReadWorker")]
-struct ReadInput {
-    downstream: EchoesOut,
-}
-
-impl ReadInput {
-    fn new() -> Self {
-        Self {
-            downstream: EchoesOut::default(),
-        }
-    }
-
-    pub async fn read_input(&self) -> Result<Envelope, WorkerError> {
-        let mut reader = FramedRead::new(stdin(), LinesCodec::new());
-        let input = reader.next().await;
-
-        match input {
-            Some(input) => {
-                let line = input.map_err(|_| WorkerError::Retry)?;
-                info!("input '{}'", line);
-                serde_json::from_str(&line).map_err(|_| WorkerError::Retry)
-            }
-            None => Err(WorkerError::Panic), // EOF
-        }
-    }
-}
-
-#[async_trait::async_trait(?Send)]
-impl gasket::framework::Worker<ReadInput> for ReadWorker {
-    async fn bootstrap(_stage: &ReadInput) -> Result<Self, WorkerError> {
-        let worker = Self {};
-
-        Ok(worker)
-    }
-
-    async fn schedule(
-        &mut self,
-        _stage: &mut ReadInput,
-    ) -> Result<WorkSchedule<Action>, WorkerError> {
-        Ok(WorkSchedule::Unit(Action {}))
-    }
-
-    async fn execute(&mut self, _unit: &Action, stage: &mut ReadInput) -> Result<(), WorkerError> {
-        let msg = stage.read_input().await.or_panic()?;
-
-        trace!("message '{:?}'", msg);
-
-        stage
-            .downstream
-            .send(msg.into())
-            .await
-            .map_err(|_| WorkerError::Panic)?;
-        Ok(())
-    }
-}
-
-#[derive(Stage)]
-#[stage(name = "do_echo", unit = "Action", worker = "EchoWorker")]
-struct DoEcho {
-    downstream: EchoesOut,
-    upstream: EchoesIn,
-    node_id: String,
-    count: u64,
-}
-
-impl DoEcho {
-    fn new() -> Self {
-        Self {
-            upstream: EchoesIn::default(),
-            downstream: EchoesOut::default(),
-            node_id: "".to_string(),
-            count: 0,
-        }
-    }
-
-    fn handle_echo(&mut self, msg: Envelope) -> Result<Envelope, WorkerError> {
-        match msg.body {
-            Init {
-                msg_id,
-                node_id,
-                node_ids,
-            } => {
-                let body = self.init(msg_id, node_id, node_ids);
-                Ok(Envelope {
-                    src: self.node_id.clone(),
-                    dest: msg.src,
-                    body,
-                })
-            }
-            Echo { msg_id, echo } => Ok(Envelope {
-                src: self.node_id.clone(),
-                dest: msg.src,
-                body: self.echo(msg_id, echo),
-            }),
-            _ => Err(WorkerError::Panic),
-        }
-    }
-
-    fn init(&mut self, msg_id: u64, node_id: String, _node_ids: Vec<String>) -> Message {
-        self.node_id = node_id;
-        InitOk {
-            in_reply_to: msg_id,
-        }
-    }
-
-    fn echo(&mut self, msg_id: u64, echo: String) -> Message {
-        self.count += 1;
-        if self.count % 5 == 0 {
-            EchoOk {
-                msg_id: self.count,
-                in_reply_to: msg_id,
-                echo: echo.to_uppercase(),
-            }
-        } else {
-            EchoOk {
-                msg_id: self.count,
-                in_reply_to: msg_id,
-                echo,
-            }
-        }
-    }
-}
-
-struct EchoWorker {}
-
-#[async_trait::async_trait(?Send)]
-impl gasket::framework::Worker<DoEcho> for EchoWorker {
-    async fn bootstrap(_stage: &DoEcho) -> Result<Self, WorkerError> {
-        let worker = Self {};
-
-        Ok(worker)
-    }
-
-    async fn schedule(&mut self, _stage: &mut DoEcho) -> Result<WorkSchedule<Action>, WorkerError> {
-        Ok(WorkSchedule::Unit(Action {}))
-    }
-
-    async fn execute(&mut self, _unit: &Action, stage: &mut DoEcho) -> Result<(), WorkerError> {
-        let from_echo = stage.upstream.recv().await.or_panic()?;
-
-        let to_echo: Envelope = stage.handle_echo(from_echo.payload)?;
-
-        stage
-            .downstream
-            .send(to_echo.into())
-            .await
-            .map_err(|_| WorkerError::Panic)
-    }
-}
-
-#[derive(Stage)]
-#[stage(name = "write_output", unit = "Action", worker = "OutputWorker")]
-struct SendOutput {
-    upstream: EchoesIn,
-}
-
-impl SendOutput {
-    fn new() -> Self {
-        Self {
-            upstream: EchoesIn::default(),
-        }
-    }
-}
-
-struct OutputWorker {}
-
-#[async_trait::async_trait(?Send)]
-impl gasket::framework::Worker<SendOutput> for OutputWorker {
-    async fn bootstrap(_stage: &SendOutput) -> Result<Self, WorkerError> {
-        let worker = Self {};
-
-        Ok(worker)
-    }
-
-    async fn schedule(
-        &mut self,
-        _stage: &mut SendOutput,
-    ) -> Result<WorkSchedule<Action>, WorkerError> {
-        Ok(WorkSchedule::Unit(Action {}))
-    }
-
-    async fn execute(&mut self, _unit: &Action, stage: &mut SendOutput) -> Result<(), WorkerError> {
-        let from_echo = stage.upstream.recv().await.or_panic()?;
-        let out = serde_json::to_string(&from_echo.payload).map_err(|_| WorkerError::Panic)?;
-        trace!("Sending '{}'", out);
-        println!("{}", out);
-        Ok(())
     }
 }
